backend/requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0
razorpay==1.4.1
twilio==8.13.0
firebase-admin==6.3.0
redis==5.0.1
python-multipart==0.0.22
python-dotenv==1.0.0
pytest==7.4.4
pytest-asyncio==0.23.3
aiofiles==23.2.1
